backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic_settings
psycopg2-binary
python-jose[cryptography]
passlib
argon2-cffi
authlib
httpx
starlette
itsdangerous
cryptography
